Inline local control
